quickfix

+++ src/Task/I18nTask.php
@@ -188,15 +188,19 @@ class I18nTask extends TaskAbstract
      * merge new PO file and remove it
      * @param string $basePo
      * @param string $newPo
+     * @param bool $withoutRemoving
      * @return bool
      */
-    protected function addPo($basePo, $newPo)
+    protected function addPo($basePo, $newPo, $withoutRemoving=false)
     {
         if (!is_file($basePo) || !is_file($newPo))
             return false;
         
         $this->exec($this->getMsgCat($newPo, $basePo));
-        unlink($newPo);
+        
+        if (!$withoutRemoving)
+            unlink($newPo);
+        
         return true;
     }
 
@@ -216,7 +220,7 @@ class I18nTask extends TaskAbstract
                 touch($poBaseFile);
             }
 
-            $this->addPo($poBaseFile, $poFile);
+            $this->addPo($poBaseFile, $poFile, true);
 
             $files[] = $poBaseFile;
         }
